Use env lookup to determine HOME for SSH pubkey
Following the merge of
https://github.com/openstack/bifrost/commit/1236d9343ec16b4166f923bd62e3153ad86fd350,
facts are no longer gathered in the deploy-dynamic.yaml playbook. The
default value for the ssh_public_key_path variable references a host
fact (ansible_env) which is therefore now invalid.

This change uses the env lookup to determine the HOME environment
variable, since we are actually referring to the home of the user
executing Ansible rather than the remote user.

Change-Id: I3642186e188f8a880246965dbf667ada992cc135
Closes-Bug: #1665681